Removed scoped styles and simplified css in preparation for Tailwind being added
