--- src/components/Footer.vue
@@ -1,21 +1,0 @@
-<template>
-    <footer class="footer">
-      <p>If interested, please contact Elizabeth at elizabethszerba@gmail.com or at (904) 349-8769</p>
-    </footer>
-  </template>
-  
-  <style scoped>
-  .footer {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    background-color: var(--background-color);
-    color: var(--text-color);
-    text-align: center;
-    padding: 10px 10px;
-    z-index: 1000;
-    box-shadow: 0px -2px 10px rgba(0, 0, 0, 0.1);
-  }
-  </style>
-
