+++ src/components/SiteFooter.vue
@@ -1,0 +1,8 @@
+<template>
+  <footer>
+    <p>If interested, please contact Elizabeth at elizabethszerba@gmail.com or at (904) 349-8769</p>
+  </footer>
+</template>
+
+<style scoped>
+</style>
