+++ src/components/SiteHeader.vue
@@ -1,0 +1,13 @@
+<template>
+  <header>
+    <BackLink />
+    <h1>Willow's Puppies</h1>
+  </header>
+</template>
+
+<script setup>
+import BackLink from './BackLink.vue';
+</script>
+
+<style scoped>
+</style>
