--- src/style.css
@@ -1,116 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Dancing+Script:wght@400;700&display=swap');
-
-.header {
-  position: fixed;
-  top: 0;
-  width: 100%;
-  height: 50px; /* Set height to match the padding in #app */
-  z-index: 1000;
-  background-color: var(--background-color);
-  color: var(--text-color);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 1rem;
-  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-:root {
-  --background-color-light: #cabea5;
-  --background-color-dark: #201e2b;
-  --text-color-light: #213547;
-  --text-color-dark: #ffffff;
-
-  /* Initial (default) colors */
-  --background-color: var(--background-color-light);
-  --text-color: var(--text-color-light);
-
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-:root[data-theme='light'] {
-  --background-color: var(--background-color-light);
-  --text-color: var(--text-color-light);
-}
-
-:root[data-theme='dark'] {
-  --background-color: var(--background-color-dark);
-  --text-color: var(--text-color-dark);
-}
-
-body {
-  margin: 0;
-  min-width: 320px;
-  min-height: 100vh;
-  background-color: var(--background-color);
-  color: var(--text-color);
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-
-a:hover {
-  color: #535bf2;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: var(--background-color); 
-  color: var(--text-color); 
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-
-button:hover {
-  border-color: #646cff;
-}
-
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-.card {
-  padding: 2em;
-  background-color: var(--background-color); 
-  color: var(--text-color); 
-}
-
-#app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem 1rem; /* Adjust the padding */
-  text-align: center;
-  box-sizing: border-box;
-  padding-top: calc(50px + 2rem); /* Adjust space for the fixed header */
-  padding-bottom: calc(50px + 2rem); /* Adjust space for the fixed footer */
-}
-
-footer {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-  height: 50px; /* Set height to match the padding in #app */
-  z-index: 1000;
-  background-color: var(--background-color);
-  color: var(--text-color);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  box-shadow: 0px -2px 10px rgba(0, 0, 0, 0.1);
-}
